% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reggen.R
\name{mygeneric}
\alias{mygeneric}
\title{A pure registry-dispatch generic}
\usage{
mygeneric(...)
}
\arguments{
\item{...}{Generic arguments.}
}
\description{
A pure registry-dispatch generic
}
\examples{
mygeneric.character <- function(...) stop("Never Called")
mygeneric("foo")
}
