% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reggen.R
\name{use_registered}
\alias{use_registered}
\title{Pure registry dispatch}
\usage{
use_registered(gen, ...)
}
\arguments{
\item{gen}{A generic function to invoke.}

\item{...}{Arguments passed to the generic function.}
}
\description{
Pure registry dispatch
}
